Create 32. Detect a Cycle in Linked List
Taking slow and fast pointers to detect the cycle in the LL, if both the slow and fast pointers meet,
Then it proves that there is a cycle in the LL.
fast.next.next!=None because take a situation where slow is at last node and fast is at 2nd last node, in the next step, both slow and fast pointers meet but at None. This should not happen.