Update and rename 44. to 44. Minimum number of platforms required for a railway station
Sort the start and end arrays individually in ascending order. It might occur that the start and departure time of a train might be jumbled ,i.e. may be present at different indexes.
We don't care about the order because we just need one instance where 
maximum platforms are occupied. That is what matters to us.