Update and rename 1. Max Sum Subarray of size K to 3. Max Sum Subarray of size K
Take 2 variable mx and summ which help in calculating sum. To keep track of window, we need a start and end pointer. So take i and j.
As long as (j-i+1)<K, just increment j. We always add Arr[j] the end element and remove the start element Arr[i] if the size is reached(j-i+1 == k).
The maximum sum is stored in the mx variable.